python/requirements.txt
mysql-connector-python
black